+++ lib/libfdt/fdt_sw.c
@@ -72,8 +72,8 @@ static int _fdt_sw_check_header(void *fdt)
 
 static void *_fdt_grab_space(void *fdt, size_t len)
 {
-	int offset = fdt_size_dt_struct(fdt);
-	int spaceleft;
+	unsigned offset = fdt_size_dt_struct(fdt);
+	unsigned spaceleft;
 
 	spaceleft = fdt_totalsize(fdt) - fdt_off_dt_struct(fdt)
 		- fdt_size_dt_strings(fdt);
@@ -89,7 +89,7 @@ int fdt_create(void *buf, int bufsize)
 {
 	void *fdt = buf;
 
-	if (bufsize < sizeof(struct fdt_header))
+	if ((unsigned)bufsize < sizeof(struct fdt_header))
 		return -FDT_ERR_NOSPACE;
 
 	memset(buf, 0, bufsize);
@@ -171,7 +171,7 @@ static int _fdt_find_add_string(void *fdt, const char *s)
 	const char *p;
 	int strtabsize = fdt_size_dt_strings(fdt);
 	int len = strlen(s) + 1;
-	int struct_top, offset;
+	unsigned struct_top, offset;
 
 	p = _fdt_find_string(strtab - strtabsize, strtabsize, s);
 	if (p)
